remove user geometry
