+++ examples/triangle/src/main.rs
@@ -4,7 +4,8 @@ extern crate embree;
 extern crate support;
 
 use embree::{
-    BufferUsage, Device, Format, IntersectContext, RayHitStream, RayStream, TriangleMesh,
+    BufferUsage, Device, Format, GeometryKind, IntersectContext, RayHitStream, RayStream,
+    SubdivisionGeometry, TriangleMesh,
 };
 
 fn main() {
